requirements/runtime.txt
# Core 5 Repository
--find-links https://indianabuild:evhgtd55rP@repos.propylon.com/core/5/sdists/test/

# INScribe
--find-links http://igarh7inscribebuild/pypi

--trusted-host igarh7inscribebuild

# Always pick the latest models.
in1_chamber_models

# Proplyon
gasp
numbering-service
joplin
joplin-drf
gopher
htmlwraplib
metadata-manager
skippy
stomp.py
trex
poster
herodotus

# Django 
Django>=1.8,<1.9.0
django-webpack-loader==0.5.0
djangorestframework==3.3.2
django-filter==1.0.2
django-crispy-forms==1.6.1

wheel==0.26.0
Unipath
